+++ mutwo/core_converters/configurations.py
@@ -1,0 +1,9 @@
+"""Configure :mod:`mutwo.core_converters`"""
+
+DEFAULT_DURATION_TO_SEARCH_NAME = "duration"
+"""Default value for :param:`duration_to_search_name` parameter in
+:class:`mutwo.core_converters.MutwoParameterDictToDuration`"""
+
+DEFAULT_DURATION_KEYWORD_NAME = "duration"
+"""Default value for :param:`duration_keyword_name` parameter in
+:class:`mutwo.core_converters.MutwoParameterDictToDuration`"""